4.21 done

+++ part4/blog/controllers/blogs.js
@@ -35,8 +35,20 @@ blogsRouter.post('/blogs', async (request, response) => {
 })
 
 blogsRouter.delete('/blogs/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  console.log(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString()===decodedToken.id){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+  else{
+    response.status(403).json({error: 'Forbidden'})
+  }
+  
 })
 
 blogsRouter.put('/blogs/:id', async (request, response) => {
